--- a/.babelrc
+++ b/.babelrc
@@ -3,7 +3,7 @@
     [
       "@babel/preset-env",
       {
-        "debug": true,
+        "debug": false,
         "modules": false,
         "useBuiltIns": "usage",
         "corejs": {