--- a/package.json
+++ b/package.json
@@ -99,6 +99,6 @@
     "webpack-bundle-analyzer": "3.7.0",
     "webpack-cli": "3.3.11",
     "webpack-dev-server": "3.11.0",
-    "webpack-merge": "4.2.2"
+    "webpack-merge": "5.7.3"
   }
 }