--- a/package.json
+++ b/package.json
@@ -19,7 +19,7 @@
     "test": "jest",
     "webpack:dev-server": "webpack-dev-server --config=webpack/env.development.js",
     "webpack:production": "webpack --config=webpack/env.production.js",
-    "webpack:stats": "yarn run webpack:production -- --env.stats=true"
+    "webpack:stats": "yarn run webpack:production --env stats"
   },
   "husky": {
     "hooks": {
@@ -97,7 +97,7 @@
     "terser-webpack-plugin": "3.0.1",
     "webpack": "4.43.0",
     "webpack-bundle-analyzer": "3.7.0",
-    "webpack-cli": "3.3.11",
+    "webpack-cli": "4.7.0",
     "webpack-dev-server": "3.11.2",
     "webpack-merge": "5.7.3"
   }