--- a/package.json
+++ b/package.json
@@ -98,7 +98,7 @@
     "webpack": "4.43.0",
     "webpack-bundle-analyzer": "3.7.0",
     "webpack-cli": "3.3.11",
-    "webpack-dev-server": "3.11.0",
+    "webpack-dev-server": "3.11.2",
     "webpack-merge": "5.7.3"
   }
 }