--- a/src/components/outsiders-mark/stylesheets/index.scss
+++ b/src/components/outsiders-mark/stylesheets/index.scss
@@ -3,7 +3,7 @@
   $image-width: 488px;
 
   &::before {
-    background: url('~Images/outsiders-mark.svg') no-repeat 0 0;
+    background: url('Images/outsiders-mark.svg') no-repeat 0 0;
     background-size: cover;
     content: '';
     display: block;