--- a/webpack/base.js
+++ b/webpack/base.js
@@ -55,7 +55,7 @@ module.exports = function (environment) {
           ],
         },
         {
-          test: /\.(svg|png|jpg|jpeg|gif)$/,
+          test: /\.(svg|png|jpg|jpeg)$/,
           type: 'asset/resource',
           generator: {
             filename: 'images/[name].[hash][ext][query]',