--- a/webpack/env.development.js
+++ b/webpack/env.development.js
@@ -1,5 +1,5 @@
 require('dotenv').config();
-const webpackMerge = require('webpack-merge');
+const { merge: webpackMerge } = require('webpack-merge');
 const CopyWebpackPlugin = require('copy-webpack-plugin');
 const ReactRefreshWebpackPlugin = require('@pmmmwh/react-refresh-webpack-plugin');
 const paths = require('../support/paths');