| Stats (Before - 98.58kb) | Stats (After - 97.45kb) |
| --- | --- |
| ![before](https://user-images.githubusercontent.com/167421/118565489-29c31800-b76a-11eb-8837-107953b09a63.png) | ![after](https://user-images.githubusercontent.com/167421/118565433-1021d080-b76a-11eb-8af4-5a0e54269828.png) |

**TLDR:**
- Simpler docker configuration.
- Simpler HTML template.
- Swapping third-party plugins for newly built-in behaviour (`file-loader` & `clean-webpack-plugin` & `terser-webpack-plugin`).
- Overall simplification of the configuration for loaders, plugins and the dev server.
- Negligible file size improvement (overall ~1kb smaller).
- Disappointing speed reduction (was average of 6.63s, is now average of 14.37s).

**Detailed Breakdown:**
- Turning off babel debug.
- Using `slim` image for dev container, and continuing to use `alpine` for web container. This is because slim worked straight out of the box in terms of libraries / binaries for building imagemin dependencies. This lets us get rid of all the `apk` commands to install various libraries.
- Adding `bash` to the web container.
- Replacing the manual script, style and meta rendering blocks for template tags.
- Swapping `webpack-dev-server` calls for `webpack serve`.
- Removing `~` from CSS paths as it's now built-in of behaviour to check node modules then webpack aliases.
- Bumping all webpack related dependencies.
- Swapping `imagemin-webpack-plugin` for `image-minimizer-webpack-plugin` and related dependencies.
- Swapping `optimize-css-assets-webpack-plugin` for `css-minimizer-webpack-plugin`.
- Swapping `node-sass` for `sass`.
- Removing `clean-webpack-plugin` for `output.clean`.
- Removing `file-loader` for asset management (`asset/resource').
- Removing `moduleIds` as it's part of `production` mode now.
- Removing split groups in favour of webpack default bundling.