@anthonyhastings 2021-07-29
:tada: This PR is included in version 1.0.0 :tada:

The release is available on [GitHub release](https://github.com/anthonyhastings/dishonored2-power-calculator/releases/tag/v1.0.0)

Your **[semantic-release](https://github.com/semantic-release/semantic-release)** bot :package::rocket: